# docs/swagger.yaml
basePath: /api
definitions:
  credit-report-service_internal_apperr.ErrorBody:
    properties:
      details: {}
      error:
        type: string
      message:
        type: string
      status:
        type: integer
      timestamp:
        type: string
    type: object
  credit-report-service_internal_models.Account:
    properties:
      createdAt:
        type: string
      dateOfBirth:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      phone:
        type: string
      profileCompleted:
        type: boolean
      role:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  credit-report-service_internal_service.AuthResult:
    properties:
      account:
        $ref: '#/definitions/credit-report-service_internal_models.Account'
      expiresAt:
        type: string
      refreshToken:
        type: string
      token:
        type: string
    type: object
...
  credit-report-service_internal_service.SignupResult:
    properties:
      accountId:
        type: integer
      email:
        type: string
      message:
        type: string
    type: object
...
  internal_handler.googleLoginReq:
    properties:
      idToken:
        example: eyJhbGciOiJSUzI1NiIs...
        type: string
    type: object
  internal_handler.loginReq:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: hunter2pass
        type: string
    type: object
...
  internal_handler.resendReq:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
  internal_handler.signupReq:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: hunter2pass
        type: string
    type: object
...
  internal_handler.updateProfileReq:
    properties:
      dateOfBirth:
        example: "1815-12-10"
        type: string
      firstName:
        example: Ada
        type: string
      lastName:
        example: Lovelace
        type: string
    type: object
  internal_handler.verifyEmailReq:
    properties:
      email:
        example: user@example.com
        type: string
      otp:
        example: "123456"
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API for the credit-report Android app: email/password auth with OTP verification, JWT sessions, profile management, and credit-report CRUD.

    All routes are mounted under `/api`. Protected routes require an `Authorization: Bearer <jwt>` header obtained from login or email verification.
  title: Credit Report Service API
  version: "1.0"
paths:
...
  /auth/google:
    post:
      consumes:
      - application/json
      description: Verifies a Google ID token (issued by the Android/iOS Google Sign-In
        SDK) and returns a session JWT. On first login, creates a verified account;
        on subsequent logins, reuses the existing account. If the Google email matches
        an existing account, the Google identity is linked onto it. Requires the Web
        OAuth client ID to be configured (AUTH_GOOGLE_CLIENT_ID); otherwise returns
        503.
      parameters:
      - description: Stable per-device UUID
        in: header
        name: X-Device-Id
        type: string
      - description: Human-readable device name shown in the device list
        in: header
        name: X-Device-Name
        type: string
      - description: ios | android | web
        in: header
        name: X-Device-Platform
        type: string
      - description: JSON device description, e.g. {\
        in: header
        name: X-Device-Info
        type: string
      - description: Google ID token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.googleLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_service.AuthResult'
        "400":
          description: Validation failed (missing idToken)
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "401":
          description: Invalid Google token
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "409":
          description: Google identity already linked to another account
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "503":
          description: Google login is not configured
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      summary: Log in with Google
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: 'Verifies email + password and opens a session for the calling
        device. Returns a short-lived access JWT plus a refresh token — in the JSON
        body for mobile clients, or as an httpOnly `refresh_token` cookie when `X-Device-Platform:
        web`. Send `X-Device-Id` (a stable per-device UUID) so repeat logins update
        one entry in the device list instead of creating a new one.'
      parameters:
      - description: Stable per-device UUID
        in: header
        name: X-Device-Id
        type: string
      - description: Human-readable device name shown in the device list
        in: header
        name: X-Device-Name
        type: string
      - description: ios | android | web
        in: header
        name: X-Device-Platform
        type: string
      - description: JSON device description, e.g. {\
        in: header
        name: X-Device-Info
        type: string
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_service.AuthResult'
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "401":
          description: Invalid email or password / email not verified
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      summary: Log in with email + password
      tags:
      - auth
...
  /auth/otp/resend:
    post:
      consumes:
      - application/json
      description: Re-issues the signup OTP for an unverified email, subject to cooldown
        / send limits.
      parameters:
      - description: Email to resend to
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.resendReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\": \"Verification code re-sent\"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation failed / cooldown / send limit
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "404":
          description: No signup found for this email
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "409":
          description: Email is already verified
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      summary: Resend signup verification OTP
      tags:
      - auth
...
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a PENDING account with an unverified password identity
        and emails a verification OTP. Re-signing up for an unverified email updates
        the password and re-issues the OTP.
      parameters:
      - description: Signup credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.signupReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/credit-report-service_internal_service.SignupResult'
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      summary: Sign up with email + password
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Checks the signup OTP; on success verifies the identity, activates
        the account, and opens a session for the calling device. Token delivery follows
        the same rules as POST /auth/login.
      parameters:
      - description: Stable per-device UUID
        in: header
        name: X-Device-Id
        type: string
      - description: Human-readable device name shown in the device list
        in: header
        name: X-Device-Name
        type: string
      - description: ios | android | web
        in: header
        name: X-Device-Platform
        type: string
      - description: JSON device description, e.g. {\
        in: header
        name: X-Device-Info
        type: string
      - description: Email + OTP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.verifyEmailReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_service.AuthResult'
        "400":
          description: Wrong / expired / locked OTP
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      summary: Verify email with OTP
      tags:
      - auth
...
  /profile:
    get:
      description: Returns the full account record for the current session.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_models.Account'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get the authenticated account's profile
      tags:
      - profile
    put:
      consumes:
      - application/json
      description: Sets first/last name (required) and optional date of birth (YYYY-MM-DD),
        marking the profile step complete.
      parameters:
      - description: Profile fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_models.Account'
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update the authenticated account's profile
      tags:
      - profile
securityDefinitions:
  BearerAuth:
    description: '"Bearer <jwt>" — obtain via POST /api/auth/login or /api/auth/verify-email'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"